requirements.txt
loguru==0.7.0
Pillow==9.5.0
tqdm==4.65.0
setuptools==65.5.1
pandas==2.0.1
numpy==1.24.3
scikit-learn==1.2.2
click==8.1.3
python-dotenv==1.0.0
kaggle==1.5.13
torchsampler==0.1.2
mlflow==2.2.0
PyYAML==6.0
dvc[s3]==2.20.0
torchmetrics==0.11.4